upgrade next.js to version 15.1.6 and update TypeScript configuration; refactor page component structure
