+++ package.json
@@ -25,7 +25,7 @@
     "jsonwebtoken": "^9.0.2",
     "jwt-decode": "^3.1.2",
     "multer": "^1.4.5-lts.1",
-    "next": "^14.0.2",
+    "next": "^15.1.6",
     "next-auth": "^4.23.1",
     "openai": "^4.19.1",
     "react": "18.2.0",
